Update index.html
Fixed a bug:  I've reviewed the code and discovered that in our new helper function (which calculates the start of the pay period), if none of the paydays qualify based on the selected day the function returns a null value. This then causes arithmetic to fail in the calculation logic so nothing gets updated. I’ve added a fallback in getMostRecentPaydayDay so that if no payday is found it defaults to 1.